Move OIDC server into functional area
